Full Auto Command Complete
Refactored original code to split functionality across the folder
structure. Implemented the auto command which retrieves the KDS root key
and the gMSA data needed to generate the gMSA key. This leverages LDAP
and assumes that the user has DA credentials. At this time the program
properly generates NTLM and AES keys that successfully authenticate to
the domain.